Refactor action images to its own component
