+++ packages/WillowCreekApp/src/ui/ActionListCard/ActionListItem.js
@@ -2,20 +2,19 @@ import React from 'react';
 import { View } from 'react-native';
 import { get } from 'lodash';
 import PropTypes from 'prop-types';
-import moment from 'moment';
 
 import {
   H6,
-  H3,
   H4,
   styled,
   TouchableScale,
-  ConnectedImage,
-  // ImageSourceType,
+  ImageSourceType,
   FlexedView,
 } from '@apollosproject/ui-kit';
 
-const StyledH6 = styled(({ theme }) => ({
+import ActionListImage from './ActionListImage';
+
+export const StyledH6 = styled(({ theme }) => ({
   color: theme.colors.text.tertiary,
 }))(H6);
 
@@ -27,7 +26,7 @@ const TextContainer = styled(({ theme }) => ({
   borderColor: theme.colors.shadows.default,
 }))(FlexedView);
 
-const Cell = styled(({ theme }) => ({
+export const Cell = styled(({ theme }) => ({
   paddingHorizontal: theme.sizing.baseUnit,
   paddingVertical: theme.sizing.baseUnit / 4,
   backgroundColor: theme.colors.background.paper,
@@ -36,45 +35,6 @@ const Cell = styled(({ theme }) => ({
   alignItems: 'center',
 }))(View);
 
-const CellImage = styled(({ theme }) => ({
-  width: theme.sizing.baseUnit * 4,
-  height: theme.sizing.baseUnit * 4,
-  borderRadius: theme.sizing.baseUnit,
-  marginRight: theme.sizing.baseUnit,
-}))(ConnectedImage);
-
-const CellView = styled(({ theme }) => ({
-  width: theme.sizing.baseUnit * 4,
-  height: theme.sizing.baseUnit * 4,
-  borderRadius: theme.sizing.baseUnit,
-  marginRight: theme.sizing.baseUnit,
-  backgroundColor: '#F4F4F5',
-  alignItems: 'center',
-  justifyContent: 'center',
-}))(View);
-
-const CellDate = styled(({ theme }) => ({
-  color: theme.colors.darkSecondary,
-  marginBottom: -theme.sizing.baseUnit / 2,
-}))(H3);
-
-const CellMonth = styled(({ theme }) => ({
-  color: theme.colors.darkSecondary,
-}))(H6);
-
-const Image = ({ relatedNode, source }) => {
-  if (source === '' && get(relatedNode, '__typename') === 'Event') {
-    const date = moment(get(relatedNode, 'start'));
-    return (
-      <CellView>
-        <CellDate>{date.format('D')}</CellDate>
-        <CellMonth>{date.format('MMM')}</CellMonth>
-      </CellView>
-    );
-  }
-  return <CellImage source={source} />;
-};
-
 const ActionListItem = ({
   imageSource,
   title,
@@ -82,22 +42,30 @@ const ActionListItem = ({
   relatedNode,
   label,
   onPressActionItem,
-}) => (
-  <TouchableScale onPress={() => onPressActionItem({ action, relatedNode })}>
-    <Cell>
-      <Image source={imageSource} relatedNode={relatedNode} />
-      <TextContainer>
-        {label ? <StyledH6 numberOfLines={1}>{label}</StyledH6> : null}
-        <H4 numberOfLines={2} ellipsizeMode="tail">
-          {title}
-        </H4>
-      </TextContainer>
-    </Cell>
-  </TouchableScale>
-);
+}) => {
+  const type = get(relatedNode, '__typename');
+  const start = get(relatedNode, 'start');
+
+  return (
+    <TouchableScale onPress={() => onPressActionItem({ action, relatedNode })}>
+      <Cell>
+        <ActionListImage source={imageSource} type={type} start={start} />
+        <TextContainer>
+          {label ? <StyledH6 numberOfLines={1}>{label}</StyledH6> : null}
+          <H4 numberOfLines={2} ellipsizeMode="tail">
+            {title}
+          </H4>
+        </TextContainer>
+      </Cell>
+    </TouchableScale>
+  );
+};
 
 ActionListItem.propTypes = {
-  // imageSource: ImageSourceType.isRequired,
+  imageSource: PropTypes.oneOfType([
+    PropTypes.arrayOf(ImageSourceType),
+    ImageSourceType,
+  ]),
   title: PropTypes.string.isRequired,
   action: PropTypes.string,
   relatedNode: PropTypes.any, // eslint-disable-line
